Added functionality to stop active visit #153
Addded feature according to issue #153
Hello @team,
We were able to add this new feature as instructed. The feature was to add a button that can stop an active patient visit. 
GROUP MEMBERS:
INTISAAR ABDUL OTHMAN | 2020-04-10222 | BSc in CS
JORAM ALLAN KINGU | 2020-04-04092 | BSc in CS
ASYA OTHMAN IBRAHIM | 2020-04-02408 | BSc in CS
STELLA JAMES HANGI | 2020-04-02198 | BSc in CS
MUDATHIR ALAWI SAID | 2020-04-10908 | BSc in CS
Feature Tracing and placement
We were able to track where the feature buttons and codes should be placed by inspecting elements from the chrome browser in order to identify which element section shows when you select active visits. Then through that inspection we were able to trace the element that contained the data of active patient and add a button at the bottom that stop active patient.
Challenges 
    • @angular/material failed to install components from the project which causes some of the pages not look good.
Adding the feature
Once the page that show patient data when active patient is clicked is found then we add new button at the bottom to stop active patient. The button is used to call the API found in openmrs in the visits category named ‘update visit’ which set the closing date of a particular patient to the current date and time. So once clicked it takes current date automatically and put it in a selected patient closing date thus ending their active visits.
What we have done is adding a closing button and the function to handle visit closing as shown below.